test_gemini.py
from dotenv import load_dotenv
import os
from groq import Groq

load_dotenv()

client = Groq(api_key=os.getenv("GROQ_API_KEY"))

response = client.chat.completions.create(
    model="llama-3.3-70b-versatile",
    messages=[{"role": "user", "content": "Say 'Groq API connected successfully' in one sentence."}],
    max_tokens=50,
)
print(response.choices[0].message.content)
