requirements.txt
yfinance>=0.2.50
pandas>=2.0.0
ta>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.11.0
streamlit>=1.40.0
plotly>=5.20.0
rich>=13.0.0
beautifulsoup4>=4.12.0
